// src/types/ui/enums/Operation.ts
export enum Operation {
	INSERT_BACK = "insert_back",
	INSERT_FRONT = "insert_front",
	INSERT_AFTER = "insert_after",
	INSERT_BEFORE = "insert_before",
	MOVE_BACK = "move_back",
	MOVE_FRONT = "move_front",
	MOVE_AFTER = "move_after",
	MOVE_BEFORE = "move_before",
	SWAP = "swap",
	REMOVE = "remove",
	REPLACE = "replace",
}